+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+### Add
+* debug logging for metadata in upload
+
+
 ## [0.3.10] - 2019-06-18
 
 ### Add